# dormitory-backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dormitory API",
    "description": "API for managing dormitory operations including authentication, admin, areas, registrations, reports, rooms, users, contracts, bills, notifications, services, and payments",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/contracts": {
      "get": {
        "tags": ["Contract"],
        "summary": "Get all contracts (Admin)",
        "description": "Retrieve a paginated list of contracts with optional filters. Any provided filters (user_id, email, room_id, status, start_date, end_date, contract_type) must have valid non-empty values.",
        "security": [{"Bearer": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}, "description": "Page number", "example": 1},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 10}, "description": "Items per page", "example": 10},
          {"name": "user_id", "in": "query", "schema": {"type": "integer"}, "description": "Filter by user ID (mutually exclusive with email)", "example": 1},
          {"name": "email", "in": "query", "schema": {"type": "string"}, "description": "Filter by user email (mutually exclusive with user_id)", "example": "user@example.com"},
          {"name": "room_id", "in": "query", "schema": {"type": "integer"}, "description": "Filter by room ID", "example": 1},
          {"name": "status", "in": "query", "schema": {"type": "string", "enum": ["PENDING", "ACTIVE", "EXPIRED", "TERMINATED"]}, "description": "Filter by status", "example": "PENDING"},
          {"name": "start_date", "in": "query", "schema": {"type": "string", "format": "date"}, "description": "Filter by start date (YYYY-MM-DD)", "example": "2025-01-01"},
          {"name": "end_date", "in": "query", "schema": {"type": "string", "format": "date"}, "description": "Filter by end date (YYYY-MM-DD)", "example": "2025-12-31"},
          {"name": "contract_type", "in": "query", "schema": {"type": "string", "enum": ["SHORT_TERM", "LONG_TERM"]}, "description": "Filter by contract type", "example": "LONG_TERM"}
          
        ],
        "responses": {
          "200": {
            "description": "List of contracts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contracts": {"type": "array", "items": {"$ref": "#/components/schemas/Contract"}},
                    "total": {"type": "integer", "example": 100},
                    "pages": {"type": "integer", "example": 10},
                    "current_page": {"type": "integer", "example": 1}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or empty query parameters",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"},
                "examples": {
                  "empty_param": {
                    "value": {"message": "Tham số status không được để trống"}
                  },
                  "invalid_status": {
                    "value": {"message": "Trạng thái hợp đồng không hợp lệ"}
                  },
                  "invalid_date": {
                    "value": {"message": "Định dạng start_date không hợp lệ (YYYY-MM-DD)"}
                  },
                  "invalid_type": {
                    "value": {"message": "Loại hợp đồng không hợp lệ"}
                  },
                  "both_user_id_and_email": {
                    "value": {"message": "Không thể cung cấp cả user_id và email cùng lúc"}
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"},
                "example": {"message": "Bạn không có quyền truy cập"}
              }
            }
          },
          "404": {
            "description": "User or room not found",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"},
                "examples": {
                  "user_not_found": {
                    "value": {"message": "Không tìm thấy người dùng"}
                  },
                  "email_not_found": {
                    "value": {"message": "Không tìm thấy người dùng với email này"}
                  },
                  "room_not_found": {
                    "value": {"message": "Không tìm thấy phòng"}
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"},
                "example": {"message": "Lỗi server khi lấy danh sách hợp đồng"}
              }
            }
          }
        }
      }
    },
    "/api/contracts/{contract_id}": {
      "get": {
        "tags": ["Contract"],
        "summary": "Get contract by ID (Admin or owner)",
        "security": [{"Bearer": []}],
        "parameters": [
          {"name": "contract_id", "in": "path", "required": true, "schema": {"type": "integer"}, "description": "Contract ID", "example": 1}
        ],
        "responses": {
          "200": {
            "description": "Contract details",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Contract"}
              }
            }
          },
          "403": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"},
                "example": {"message": "Bạn không có quyền xem hợp đồng này"}
              }
            }
          },
          "404": {
            "description": "Contract not found",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"},
                "example": {"message": "Không tìm thấy hợp đồng"}
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"},
                "example": {"message": "Lỗi server khi lấy chi tiết hợp đồng"}
              }
            }
          }
        }
      }
    },
...
    "/api/me/contracts": {
      "get": {
        "tags": ["Contract"],
        "summary": "Get current user's contracts",
        "security": [{"Bearer": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}, "description": "Page number", "example": 1},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 10}, "description": "Items per page", "example": 10}
        
        ],
        "responses": {
          "200": {
            "description": "List of user's contracts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contracts": {"type": "array", "items": {"$参": "#/components/schemas/Contract"}},
                    "total": {"type": "integer", "example": 50},
                    "pages": {"type": "integer", "example": 5},
                    "current_page": {"type": "integer", "example": 1}
                  }
                }
              }
            }
          },
          "403": {
            "description": "User access required",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"},
                "example": {"message": "Bạn không có quyền truy cập"}
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Error"},
                "example": {"message": "Lỗi server khi lấy danh sách hợp đồng của người dùng"}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Contract": {
        "type": "object",
        "properties": {
          "contract_id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique ID of the contract",
            "example": 1
          },
          "room_id": {
            "type": "integer",
            "description": "ID of the room",
            "example": 1
          },
          "user_id": {
            "type": "integer",
            "format": "int64",
            "description": "ID of the user",
            "example": 1
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "ACTIVE", "EXPIRED", "TERMINATED"],
            "description": "Status of the contract",
            "example": "PENDING"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp",
            "example": "2025-04-14T10:00:00Z"
          },
          "contract_type": {
            "type": "string",
            "enum": ["SHORT_TERM", "LONG_TERM"],
            "description": "Type of contract",
            "example": "LONG_TERM"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "Start date of the contract",
            "example": "2025-05-01"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "description": "End date of the contract",
            "example": "2025-12-31"
          },
          "room_details": {
            "type": "object",
            "nullable": true,
            "description": "Details of the room (structure depends on Room model)"
          },
          "user_details": {
            "type": "object",
            "nullable": true,
            "description": "Details of the user (structure depends on User model)"
          }
        },
        "required": [
          "contract_id",
          "room_id",
          "user_id",
          "status",
          "created_at",
          "contract_type",
          "start_date",
          "end_date"
        ]
      }
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Không tìm thấy hợp đồng"
          }
        },
        "required": ["message"]
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}